Add unique filenames
Previously, the filename was just the main-text, meaning covers with the
same main-text would overwrite each other. I've changed the code so that
other config items and numerical suffixes will be used to build a unique
filename.

This created an issue where running the project on an existing
covers directory would create copies as the files already existed. I've
added some code then to, as well as check if the files exist, check if
the files have been modified since the start of the run. If they have,
then that's an existing filename. If they're older, they can be
overwritten. This means if you move the order of cover definitions,
then filenames may change, so at your own risk (not that I'm typing
this to anyone except me)